// src/main/model/SortType.enum.ts
export enum SortType {
    NAME = "name",
    MAIN = "main",
    MAINP = "mainp",
    COMP = "comp",
    AVERAGEA = "averagea",
    RATING = "rating",
    POPULAR = "popular",
    BACKLOG = "backlog",
    USERSP = "usersp",
    PLAYING = "playing",
    SPEEDRUNS = "speedruns",
    RELEASE = "release"
  }