// src/main/model/GamePlatforms.enum.ts
export enum GamePlatform {
  ALL = "",
  THREE_DO = "3DO",
  AMIGA = "Amiga",
  AMSTRAD_CPC = "Amstrad CPC",
  ANDROID = "Android",
  APPLE_II = "Apple II",
  ARCADE = "Arcade",
  ATARI_2600 = "Atari 2600",
  ATARI_5200 = "Atari 5200",
  ATARI_7800 = "Atari 7800",
  ATARI_8_BIT_FAMILY = "Atari 8-bit Family",
  ATARI_JAGUAR = "Atari Jaguar",
  ATARI_JAGUAR_CD = "Atari Jaguar CD",
  ATARI_LYNX = "Atari Lynx",
  ATARI_ST = "Atari ST",
  BBC_MICRO = "BBC Micro",
  BROWSER = "Browser",
  COLECOVISION = "ColecoVision",
  COMMODORE_64 = "Commodore 64",
  DREAMCAST = "Dreamcast",
  EMULATED = "Emulated",
  FM_TOWNS = "FM Towns",
  GAME_AND_WATCH = "Game & Watch",
  GAME_BOY = "Game Boy",
  GAME_BOY_ADVANCE = "Game Boy Advance",
  GAME_BOY_COLOR = "Game Boy Color",
  GEAR_VR = "Gear VR",
  HTC_VIVE = "HTC Vive",
  INTELLIVISION = "Intellivision",
  INTERACTIVE_MOVIE = "Interactive Movie",
  IOS = "iOS",
  LINUX = "Linux",
  MAC = "Mac",
  MOBILE = "Mobile",
  MSX = "MSX",
  N_GAGE = "N-Gage",
  NEO_GEO = "Neo Geo",
  NEO_GEO_CD = "Neo Geo CD",
  NEO_GEO_POCKET = "Neo Geo Pocket",
  NEO_GEO_POCKET_COLOR = "Neo Geo Pocket Color",
  NES = "NES",
  NINTENDO_2DS = "Nintendo 2DS",
  NINTENDO_3DS = "Nintendo 3DS",
  NINTENDO_64 = "Nintendo 64",
  NINTENDO_DS = "Nintendo DS",
  NINTENDO_GAMECUBE = "Nintendo GameCube",
  NINTENDO_SWITCH = "Nintendo Switch",
  OCULUS_GO = "Oculus Go",
  OCULUS_RIFT = "Oculus Rift",
  ONLIVE = "OnLive",
  OUYA = "Ouya",
  PC = "PC",
  PHILIPS_CD_I = "Philips CD-i",
  PHILIPS_VIDEOPAC_G7000 = "Philips Videopac G7000",
  PLAYSTATION = "PlayStation",
  PLAYSTATION_2 = "PlayStation 2",
  PLAYSTATION_3 = "PlayStation 3",
  PLAYSTATION_4 = "PlayStation 4",
  PLAYSTATION_MOBILE = "PlayStation Mobile",
  PLAYSTATION_NOW = "PlayStation Now",
  PLAYSTATION_VITA = "PlayStation Vita",
  PLAYSTATION_VR = "PlayStation VR",
  PLUG_AND_PLAY = "Plug & Play",
  PSP = "PSP",
  SEGA_32X = "Sega 32X",
  SEGA_CD = "Sega CD",
  SEGA_GAME_GEAR = "Sega Game Gear",
  SEGA_MASTER_SYSTEM = "Sega Master System",
  SEGA_MEGA_DRIVE_GENESIS = "Sega Mega Drive/Genesis",
  SEGA_SATURN = "Sega Saturn",
  SHARP_X68000 = "Sharp X68000",
  SUPER_NINTENDO = "Super Nintendo",
  TIGER_HANDHELD = "Tiger Handheld",
  TURBOGRAFX_16 = "TurboGrafx-16",
  TURBOGRAFX_CD = "TurboGrafx-CD",
  VIRTUAL_BOY = "Virtual Boy",
  WII = "Wii",
  WII_U = "Wii U",
  WINDOWS_PHONE = "Windows Phone",
  WONDERSWAN = "WonderSwan",
  WONDERSWAN_COLOR = "WonderSwan Color",
  XBOX = "Xbox",
  XBOX_360 = "Xbox 360",
  XBOX_ONE = "Xbox One",
  ZX_SPECTRUM = "ZX Spectrum"
}